requirements.txt
echo "Django==4.2
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.0
stripe==7.0.0
django-allauth==0.57.0" > requirements.txt
Pillow==10.0.0
